+++ terratex_reallife/ADMINSYSTEM/dev_server_connect_control.lua
@@ -13,7 +13,7 @@ local deleteNullAfterDays=30
 --[[ SCRIPT ]]
 function onPlayerDevServerConnect(nickname)
     if(fileExists(":"..getResourceName(getThisResource()).."/devmode.dev")) then
-        local isDev = MySql.helper.getFieldValueSync("players", "isDeveloper", {Nickname = nickname});
+        local isDev = MySql.helper.getValueSync("players", "isDeveloper", {Nickname = nickname});
         if(isDev)then
             if(isDev==0)then
                 local timeQuery="Nickname='"..nicknameSecure.."' and TIMEDIFF(fromTimestamp, NOW())<0 and (TIMEDIFF(toTimestamp, NOW())>0 or toTimestamp IS NULL)"