+++ terratex_reallife/ADMINSYSTEM/punishchecker.lua
@@ -30,7 +30,7 @@ function checkAdditionalPunishment_offline(playerName)
                     save_offline_message(dasatz["Nickname"], "Bewährungssystem", "Durch dein Fehlverhalten wurde deine Bewährungsstrafe nun durchgeführt und du hast " .. dasatz["warn"] .. " Warns erhalten")
 
                     if (MySQL_GetResultsCount("warns", "Nickname='" .. dasatz["Nickname"] .. "'") > 2) then
-                        local serial = MySql.helper.getFieldValueSync("players", "Serial", { Nickname = dasatz["Nickname"] });
+                        local serial = MySql.helper.getValueSync("players", "Serial", { Nickname = dasatz["Nickname"] });
 
                         local qquery = "INSERT INTO ban (Nickname,Serial,IP,Grund,Admin) VALUES ('" .. dasatz["Nickname"] .. "','" .. serial .. "','0','3 Warns','Warnsystem')"
                         mysql_query(handler, qquery)
@@ -44,7 +44,7 @@ function checkAdditionalPunishment_offline(playerName)
                     tryKickPlayer = true
                 end
                 if (tonumber(dasatz["perma"]) > 0) then
-                    local serial = MySql.helper.getFieldValueSync("players", "Serial", { Nickname = dasatz["Nickname"] });
+                    local serial = MySql.helper.getValueSync("players", "Serial", { Nickname = dasatz["Nickname"] });
                     local qquery = "INSERT INTO ban (Nickname,Serial,IP,Grund,Admin) VALUES ('" .. dasatz["Nickname"] .. "','" .. serial .. "','0','Bewährungsstrafe: " .. dasatz["Grund"] .. "','" .. dasatz["Adminname"] .. "')"
                     mysql_query(handler, qquery)
                     tryKickPlayer = true