+++ terratex_reallife/ENVIRO/fraktionen/fraktion.lua
@@ -70,120 +70,120 @@ function teamserstellen()
     --Set Settings of the Fraktion
     --SAPD
     vioSetElementData(team[1], "hidden", false)
-    frakkasse[1] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 1 });
+    frakkasse[1] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 1 });
     blacklist[1] = false;
     frakmun[1] = false;
     frakdrogen[1] = false;
     frakschutz[1] = false;
-    fraksteuer[1] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 1 });
+    fraksteuer[1] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 1 });
 
     -- Grove Street
     vioSetElementData(team[2], "hidden", false)
     blacklist[2] = {};
-    frakkasse[2] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 2 });
-    frakmun[2] = MySql.helper.getFieldValueSync("fraktionskasse", "Munition", { FrakID = 2 });
-    frakdrogen[2] = MySql.helper.getFieldValueSync("fraktionskasse", "Drogen", { FrakID = 2 });
+    frakkasse[2] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 2 });
+    frakmun[2] = MySql.helper.getValueSync("fraktionskasse", "Munition", { FrakID = 2 });
+    frakdrogen[2] = MySql.helper.getValueSync("fraktionskasse", "Drogen", { FrakID = 2 });
     frakschutz[2] = {};
-    fraksteuer[2] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 2 });
+    fraksteuer[2] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 2 });
 
     --NEWS
     vioSetElementData(team[3], "hidden", false)
     blacklist[3] = false;
-    frakkasse[3] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 3 });
+    frakkasse[3] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 3 });
     frakmun[3] = false;
     frakdrogen[3] = false;
     frakschutz[3] = false;
-    fraksteuer[3] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 3 });
+    fraksteuer[3] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 3 });
 
     --TAXI
     vioSetElementData(team[4], "hidden", false)
     blacklist[4] = false;
-    frakkasse[4] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 4 });
+    frakkasse[4] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 4 });
     frakmun[4] = false;
     frakdrogen[4] = false;
     frakschutz[4] = false;
-    fraksteuer[4] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 4 });
+    fraksteuer[4] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 4 });
 
     -- SEK
     vioSetElementData(team[5], "hidden", false)
     blacklist[5] = false;
-    frakkasse[5] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 5 });
+    frakkasse[5] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 5 });
     frakmun[5] = false;
     frakdrogen[5] = false;
     frakschutz[5] = false;
-    fraksteuer[5] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 5 });
+    fraksteuer[5] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 5 });
 
     -- LLS
     vioSetElementData(team[6], "hidden", false)
     blacklist[6] = {};
-    frakkasse[6] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 6 });
-    frakmun[6] = MySql.helper.getFieldValueSync("fraktionskasse", "Munition", { FrakID = 6 });
-    frakdrogen[6] = MySql.helper.getFieldValueSync("fraktionskasse", "Drogen", { FrakID = 6 });
+    frakkasse[6] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 6 });
+    frakmun[6] = MySql.helper.getValueSync("fraktionskasse", "Munition", { FrakID = 6 });
+    frakdrogen[6] = MySql.helper.getValueSync("fraktionskasse", "Drogen", { FrakID = 6 });
     frakschutz[6] = {};
-    fraksteuer[6] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 6 });
+    fraksteuer[6] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 6 });
 
     -- NAME
     vioSetElementData(team[7], "hidden", false)
     blacklist[7] = false;
-    frakkasse[7] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 7 });
+    frakkasse[7] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 7 });
     frakmun[7] = false;
     frakdrogen[7] = false;
     frakschutz[7] = false;
-    fraksteuer[7] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 7 });
+    fraksteuer[7] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 7 });
 
     -- Black Aliens
     vioSetElementData(team[8], "hidden", true)
     blacklist[8] = false;
-    frakkasse[8] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 8 });
+    frakkasse[8] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 8 });
     frakmun[8] = false;
     frakdrogen[8] = false;
     frakschutz[8] = false;
-    fraksteuer[8] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 8 });
+    fraksteuer[8] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 8 });
 
     --LVPD
     vioSetElementData(team[9], "hidden", true)
     blacklist[9] = false;
-    frakkasse[9] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 9 });
+    frakkasse[9] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 9 });
     frakmun[9] = false;
     frakdrogen[9] = false;
     frakschutz[9] = false;
-    fraksteuer[9] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 9 });
+    fraksteuer[9] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 9 });
 
     -- Medics
     vioSetElementData(team[10], "hidden", false)
     blacklist[10] = false;
-    frakkasse[10] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 10 });
+    frakkasse[10] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 10 });
     frakmun[10] = false;
     frakdrogen[10] = false;
     frakschutz[10] = false;
-    fraksteuer[10] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 10 });
+    fraksteuer[10] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 10 });
 
     -- CashForce
     vioSetElementData(team[11], "hidden", false)
     blacklist[11] = {};
-    frakkasse[11] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 11 });
-    frakmun[11] = MySql.helper.getFieldValueSync("fraktionskasse", "Munition", { FrakID = 11 });
-    frakdrogen[11] = MySql.helper.getFieldValueSync("fraktionskasse", "Drogen", { FrakID = 11 });
+    frakkasse[11] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 11 });
+    frakmun[11] = MySql.helper.getValueSync("fraktionskasse", "Munition", { FrakID = 11 });
+    frakdrogen[11] = MySql.helper.getValueSync("fraktionskasse", "Drogen", { FrakID = 11 });
     frakschutz[11] = {};
-    fraksteuer[11] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 11 });
+    fraksteuer[11] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 11 });
 
     -- Violettas
     vioSetElementData(team[12], "hidden", false)
     blacklist[12] = {};
-    frakkasse[12] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 12 });
-    frakmun[12] = MySql.helper.getFieldValueSync("fraktionskasse", "Munition", { FrakID = 12 });
-    frakdrogen[12] = MySql.helper.getFieldValueSync("fraktionskasse", "Drogen", { FrakID = 12 });
+    frakkasse[12] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 12 });
+    frakmun[12] = MySql.helper.getValueSync("fraktionskasse", "Munition", { FrakID = 12 });
+    frakdrogen[12] = MySql.helper.getValueSync("fraktionskasse", "Drogen", { FrakID = 12 });
     frakschutz[12] = {};
-    fraksteuer[12] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 12 });
+    fraksteuer[12] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 12 });
 
     --Racers
     vioSetElementData(team[13], "hidden", false)
     blacklist[13] = {};
-    frakkasse[13] = MySql.helper.getFieldValueSync("fraktionskasse", "Kasse", { FrakID = 13 });
-    frakmun[13] = MySql.helper.getFieldValueSync("fraktionskasse", "Munition", { FrakID = 13 });
-    frakdrogen[13] = MySql.helper.getFieldValueSync("fraktionskasse", "Drogen", { FrakID = 13 });
+    frakkasse[13] = MySql.helper.getValueSync("fraktionskasse", "Kasse", { FrakID = 13 });
+    frakmun[13] = MySql.helper.getValueSync("fraktionskasse", "Munition", { FrakID = 13 });
+    frakdrogen[13] = MySql.helper.getValueSync("fraktionskasse", "Drogen", { FrakID = 13 });
     frakschutz[13] = {};
-    fraksteuer[13] = MySql.helper.getFieldValueSync("fraktionskasse", "Steuersatz", { FrakID = 13 });
+    fraksteuer[13] = MySql.helper.getValueSync("fraktionskasse", "Steuersatz", { FrakID = 13 });
 end
 
 addEventHandler("onResourceStart", getResourceRootElement(getThisResource()), teamserstellen)
@@ -531,7 +531,7 @@ function setblacklist_func(thePlayer, command, toPlayerName, grund, ...)
                                 local resultre = mysql_query(handler, query)
                                 mysql_free_result(resultre)
 
-                                local id = MySql.helper.getFieldValueSync("blacklist", "ID", {Name = toPlayerName, Fraktion = vioGetElementData(thePlayer, "fraktion")});
+                                local id = MySql.helper.getValueSync("blacklist", "ID", {Name = toPlayerName, Fraktion = vioGetElementData(thePlayer, "fraktion")});
 
                                 table.insert(blacklist[vioGetElementData(thePlayer, "fraktion")], { toPlayerName, id })
                                 showError(thePlayer, "Dieser Spieler ist jetzt auf der Blacklist!")