+++ terratex_reallife/ENVIRO/premium/premium.lua
@@ -1,8 +1,8 @@
 function premium_func(thePlayer)
 	local time=getRealTime()
 	local nickname=getPlayerName(thePlayer)
-	local premiumOutTime = MySql.helper.getFieldValueSync("premium", "PremiumUntil", {Name = nickname}) - time.timestamp;
-	local hasPremGutSchein = MySql.helper.getFieldValueSync("premium", "PremiumGutScheine", {Name = nickname});
+	local premiumOutTime = MySql.helper.getValueSync("premium", "PremiumUntil", {Name = nickname}) - time.timestamp;
+	local hasPremGutSchein = MySql.helper.getValueSync("premium", "PremiumGutScheine", {Name = nickname});
 	triggerClientEvent(thePlayer,"showPremiumInfoGUI",thePlayer,premiumOutTime,hasPremGutSchein)
 end
 addCommandHandler("premium",premium_func,false,false)