+++ terratex_reallife/ENVIRO/spawn_kill_system/spawn.lua
@@ -120,7 +120,7 @@ function setPlayerSpawn(source, spawn, skinid, fraktion, firstspawn)
 
                         local time = getRealTime()
                         local nickname = getPlayerName(source)
-                        local premiumOutTime = MySql.helper.getFieldValueSync("premium", "PremiumUntil", { Name = nickname }) - time.timestamp;
+                        local premiumOutTime = MySql.helper.getValueSync("premium", "PremiumUntil", { Name = nickname }) - time.timestamp;
                         if (premiumOutTime > 0) then
                             local slot = vioGetElementData(source, "PremiumSpawn")
                             if isElement(vioGetElementData(source, "slot" .. slot)) then
@@ -146,7 +146,7 @@ function setPlayerSpawn(source, spawn, skinid, fraktion, firstspawn)
                     elseif (spawn == 5) then
                         local time = getRealTime()
                         local nickname = getPlayerName(source)
-                        local premiumOutTime = MySql.helper.getFieldValueSync("premium", "PremiumUntil", { Name = nickname }) - time.timestamp;
+                        local premiumOutTime = MySql.helper.getValueSync("premium", "PremiumUntil", { Name = nickname }) - time.timestamp;
                         if (premiumOutTime > 0) then
                             if vioGetElementData(source, "job") > 0 then
                                 --outputChatBox(tostring(vioGetElementData(source,"job")))
@@ -184,13 +184,13 @@ function setPlayerSpawn(source, spawn, skinid, fraktion, firstspawn)
                     local thePlayer = source
                     vioSetElementData(source, "firstspawn", false)
                     local nickname = getPlayerName(thePlayer)
-                    local hp = MySql.helper.getFieldValueSync("inventar", "Offline_HP", { Nickname = nickname });
-                    local armor = MySql.helper.getFieldValueSync("inventar", "Offline_Armor", { Nickname = nickname });
-                    local weapons = MySql.helper.getFieldValueSync("inventar", "Offline_Weapons", { Nickname = nickname });
-                    local ammo = MySql.helper.getFieldValueSync("inventar", "Offline_Ammo", { Nickname = nickname });
+                    local hp = MySql.helper.getValueSync("inventar", "Offline_HP", { Nickname = nickname });
+                    local armor = MySql.helper.getValueSync("inventar", "Offline_Armor", { Nickname = nickname });
+                    local weapons = MySql.helper.getValueSync("inventar", "Offline_Weapons", { Nickname = nickname });
+                    local ammo = MySql.helper.getValueSync("inventar", "Offline_Ammo", { Nickname = nickname });
                     local source = thePlayer
 
-                    local resetweapons = MySql.helper.getFieldValueSync("userdata", "resetWeaponsAtNextLogin", { Nickname = nickname });
+                    local resetweapons = MySql.helper.getValueSync("userdata", "resetWeaponsAtNextLogin", { Nickname = nickname });
                     if (tonumber(hp) > 0) then
                         setElementHealth(source, hp)
                     end
@@ -266,7 +266,7 @@ function setPlayerSpawnWeapons(thePlayer, Fraktion, Rang)
         end
     end
     local nickname = getPlayerName(thePlayer)
-    local resetweapons = MySql.helper.getFieldValueSync("userdata", "resetWeaponsAtNextLogin", { Nickname = nickname });
+    local resetweapons = MySql.helper.getValueSync("userdata", "resetWeaponsAtNextLogin", { Nickname = nickname });
     if (resetweapons == 1) then
         takeAllWeapons(thePlayer)
         MySQL_SetVar("userdata", "resetWeaponsAtNextLogin", 0, "Nickname='" .. nickname .. "'")