+++ terratex_reallife/SYSTEM/vehsys/privCarCommands.lua
@@ -94,10 +94,10 @@ function reloadCarsDB_func(thePlayer, Command, toPlayerName)
                     changedCars = changedCars + 1
                     local ele = theVehicle[3]
                     local id = vioGetElementData(ele, "dbid")
-                    local newmodel = MySql.helper.getFieldValueSync("vehicles", "Model", { ID = id });
+                    local newmodel = MySql.helper.getValueSync("vehicles", "Model", { ID = id });
 
                     setElementModel(ele, newmodel)
-                    local colorstr = MySql.helper.getFieldValueSync("vehicles", "Colors", { ID = id });
+                    local colorstr = MySql.helper.getValueSync("vehicles", "Colors", { ID = id });
                     vioSetElementData(ele, "colors", colorstr)
                     local colors = {}
                     local counter = 0
@@ -113,13 +113,13 @@ function reloadCarsDB_func(thePlayer, Command, toPlayerName)
                         counter = counter + 1
                     end
                     setVehicleColor(theVehicle, tonumber(colors[0]), tonumber(colors[1]), tonumber(colors[2]), tonumber(colors[3]))
-                    local tuning = getStringComponents(MySql.helper.getFieldValueSync("vehicles", "Tuning", { ID = id }))
+                    local tuning = getStringComponents(MySql.helper.getValueSync("vehicles", "Tuning", { ID = id }))
                     for theKey, theTuning in ipairs(tuning) do
                         if (theTuning ~= 0) then
                             addVehicleUpgrade(theVehicle, theTuning)
                         end
                     end
-                    local no_handel = MySql.helper.getFieldValueSync("vehicles", "no_handel", { ID = id });
+                    local no_handel = MySql.helper.getValueSync("vehicles", "no_handel", { ID = id });
 
                     vioSetElementData(thevehicle, "model", newmodel)
                     vioSetElementData(thevehicle, "colors", colorstr)